+++ src/retaggr/engines/saucenao/handlers/__init__.py
@@ -1,4 +1,5 @@
 from .danbooru import DanbooruHandler
+from .gelbooru import GelbooruHandler
 from .e621 import E621Handler
 from .konachan import KonachanHandler
 from .yandere import YandereHandler